BotRead/requirements.txt
aiogram==2.9
aiohttp==3.6.2
aioredis==1.3.1
asyncpg==0.21.0
python-dotenv
lxml==4.6.2
bs4==0.0.1
beautifulsoup4==4.9.3
requests==2.25.0
SQLAlchemy==1.3.20
gino